[tests] Rename 'TestCase' classes to 'TestCaseTest' to conform with test style
