+++ src/test/java/org/mybatis/guice/MyBatisModuleTestCaseTest.java
@@ -15,15 +15,16 @@
  */
 package org.mybatis.guice;
 
+import java.util.List;
+
 import javax.inject.Inject;
 import javax.inject.Named;
 
-import org.apache.ibatis.session.SqlSessionFactory;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-@ExtendWith(XMLGuiceTestExtension.class)
-public final class XMLMyBatisModuleTestCase extends AbstractMyBatisModuleTestCase {
+@ExtendWith(GuiceTestExtension.class)
+final class MyBatisModuleTestCaseTest extends AbstractMyBatisModuleTestCase {
 
   @Inject
   @Named("contactWithAddress")
@@ -34,14 +35,6 @@ public final class XMLMyBatisModuleTestCase extends AbstractMyBatisModuleTestCas
 
   @Inject
   private AddressConverter addressConverter;
-
-  @Inject
-  private SqlSessionFactory sqlSessionFactory;
-
-  @Test
-  void testEnvironmentId() throws Exception {
-    assert "test".equals(sqlSessionFactory.getConfiguration().getEnvironment().getId());
-  }
 
   @Test
   void testAddressConverter() throws Exception {
@@ -66,6 +59,12 @@ public final class XMLMyBatisModuleTestCase extends AbstractMyBatisModuleTestCas
     Contact contact = this.contactMapperClient.selectById(this.contactWithAdress.getId());
     assert contact != null : "impossible to retrieve Contact with id '" + this.contactWithAdress.getId() + "'";
     assert this.contactWithAdress.equals(contact) : "Expected " + this.contactWithAdress + " but found " + contact;
+  }
+
+  @Test
+  void selectAllContactsWithDatabaseId() throws Exception {
+    List<Contact> contacts = this.contactMapperClient.getAllWithDatabaseId();
+    assert contacts.size() > 0 : "Expected not empty contact table";
   }
 
 }
